TaskRestController now use @PathVariable, moved dto's to dedicated package.
