+++ src/main/java/com/example/todo_list/models/dto/RegisterDTO.java
@@ -1,5 +1,6 @@
-package com.example.todo_list.models;
+package com.example.todo_list.models.dto;
 
+import com.example.todo_list.models.UserEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 public record RegisterDTO(String username, String password) {
