+++ src/main/java/com/example/todo_list/services/TasksServiceImp.java
@@ -4,7 +4,7 @@ import com.example.todo_list.exceptions.TaskNotFoundException;
 import com.example.todo_list.exceptions.UnauthorizedException;
 import com.example.todo_list.exceptions.UserNotFoundException;
 import com.example.todo_list.models.Task;
-import com.example.todo_list.models.TaskDTO;
+import com.example.todo_list.models.dto.TaskDTO;
 import com.example.todo_list.models.TaskStatus;
 import com.example.todo_list.models.UserEntity;
 import com.example.todo_list.respositories.TaskRepository;
