+++ src/main/java/com/example/todo_list/web/rest/TaskRestController.java
@@ -1,7 +1,7 @@
 package com.example.todo_list.web.rest;
 
 import com.example.todo_list.models.Task;
-import com.example.todo_list.models.TaskDTO;
+import com.example.todo_list.models.dto.TaskDTO;
 import com.example.todo_list.services.TasksService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.security.Principal;
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/tasks")
@@ -20,12 +21,17 @@ public class TaskRestController {
     private final TasksService tasksService;
 
     @GetMapping
-    public ResponseEntity<?> task(@RequestParam(value = "taskId", required = false) Long taskId, Principal principal) {
-        if (taskId == null) {
-            log.info("about to retrieve all tasks for user [{}]", principal.getName());
+    List<TaskDTO> tasks(Principal principal) {
+        log.info("about to retrieve all tasks for user [{}]", principal.getName());
 
-            return ResponseEntity.ok(tasksService.getAllTasksByUser(principal.getName()));
-        }
+        return tasksService.getAllTasksByUser(principal.getName())
+                .stream()
+                .map(TaskDTO::from)
+                .toList();
+    }
+
+    @GetMapping("/{taskId}")
+    public ResponseEntity<?> task(@PathVariable(value = "taskId") Long taskId, Principal principal) {
         log.info("about to retrieve task with id [{}] for user [{}]", taskId, principal.getName());
 
         return ResponseEntity.ok(tasksService.getTaskById(taskId, principal.getName()));
@@ -40,26 +46,26 @@ public class TaskRestController {
         return ResponseEntity.created(URI.create("/api/tasks/" + result.getId())).build();
     }
 
-    @PutMapping
-    public ResponseEntity<TaskDTO> updateTask(@RequestParam Long taskId, @RequestBody TaskDTO taskDTO, Principal principal) {
+    @PutMapping("/{taskId}")
+    public ResponseEntity<TaskDTO> updateTask(@PathVariable Long taskId, @RequestBody TaskDTO taskDTO, Principal principal) {
         log.info("about to update task [{}], new data [{}] for user [{}]", taskId, taskDTO, principal.getName());
 
         Task result = tasksService.editTask(taskId, taskDTO, principal.getName());
 
-        return ResponseEntity.ok(TaskDTO.TaskToDTO(result));
+        return ResponseEntity.ok(TaskDTO.from(result));
     }
 
-    @PatchMapping
-    public ResponseEntity<TaskDTO> toggleStatus(@RequestParam Long taskId, Principal principal) {
+    @PatchMapping("/{taskId}")
+    public ResponseEntity<TaskDTO> toggleStatus(@PathVariable Long taskId, Principal principal) {
         log.info("about to toogle status for task [{}] by user [{}]", taskId, principal.getName());
 
         Task result = tasksService.changeTaskStatus(taskId, principal.getName());
 
-        return ResponseEntity.ok(TaskDTO.TaskToDTO(result));
+        return ResponseEntity.ok(TaskDTO.from(result));
     }
 
-    @DeleteMapping
-    public ResponseEntity<?> deleteTask(@RequestParam Long taskId, Principal principal) {
+    @DeleteMapping("/{taskId}")
+    public ResponseEntity<?> deleteTask(@PathVariable Long taskId, Principal principal) {
         log.info("about to delete task [{}] for user [{}]", taskId, principal.getName());
 
         tasksService.deleteTask(taskId, principal.getName());
